Integrate TaskList class into Usagi
