+++ src/main/java/Todo.java
@@ -1,4 +1,34 @@
-package PACKAGE_NAME;
+/**
+ * Represents a Todo task
+ */
+public class Todo extends Task {
 
-public class Todo {
+    Todo(String description) {
+        super(description);
+    }
+
+    Todo(String description, boolean isDone) {
+        super(description, isDone);
+    }
+
+    @Override
+    String getTaskType() {
+        return "[T]";
+    }
+
+    @Override
+    String getFullDescription() {
+        return getTaskType() + super.toString();
+    }
+
+    @Override
+    public String toString() {
+        return getFullDescription();
+    }
+
+    @Override
+    String toFileString() {
+        return "T | " + (isDone ? "1" : "0") + " | " + description;
+    }
+
 }
