+++ src/main/java/Ui.java
@@ -1,2 +1,3 @@
 public class Ui {
+
 }
