+++ src/main/java/Usagi.java
@@ -1,13 +1,9 @@
 import java.io.IOException;
-import java.time.format.DateTimeParseException;
 import java.util.Scanner;
 import java.util.ArrayList;
 import java.util.NoSuchElementException;
 import java.io.File;
 import java.io.FileWriter;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 
 public class Usagi {
     private static final String HORIZONTAL_LINE = "____________________________________________________________";
@@ -47,166 +43,8 @@ public class Usagi {
         }
     }
 
-    // Base Task class
-    abstract static class Task {
-        protected String description;
-        protected boolean isDone;
-
-        Task(String description) {
-            this.description = description;
-            this.isDone = false;
-        }
-
-        Task(String description, boolean isDone) {
-            this.description = description;
-            this.isDone = isDone;
-        }
-
-        String getStatusIcon() {
-            return isDone ? "[X]" : "[ ]";
-        }
-
-        void markAsDone() {
-            this.isDone = true;
-        }
-
-        void markAsNotDone() {
-            this.isDone = false;
-        }
-
-        @Override
-        public String toString() {
-            return getStatusIcon() + " " + description;
-        }
-
-        abstract String toFileString();
-        abstract String getTaskType();
-        abstract String getFullDescription();
-    }
-
-    static class Todo extends Task {
-        Todo(String description) {
-            super(description);
-        }
-
-        Todo(String description, boolean isDone) {
-            super(description, isDone);
-        }
-
-        @Override
-        String getTaskType() {
-            return "[T]";
-        }
-
-        @Override
-        String getFullDescription() {
-            return getTaskType() + super.toString();
-        }
-
-        @Override
-        public String toString() {
-            return getFullDescription();
-        }
-
-        @Override
-        String toFileString() {
-            return "T | " + (isDone ? "1" : "0") + " | " + description;
-        }
-    }
-
-    static class Deadline extends Task {
-        private static final DateTimeFormatter IO_FMT = DateTimeFormatter.ISO_LOCAL_DATE;      // e.g., 2025-08-29
-        private static final DateTimeFormatter VIEW_FMT = DateTimeFormatter.ofPattern("MMM d yyyy"); // e.g., Aug 29 2025
-
-        private final LocalDate due;
-
-        // dueDate must be ISO: yyyy-MM-dd
-        Deadline(String description, String dueDate) {
-            super(description);
-            this.due = parseDue(dueDate);
-        }
-
-        Deadline(String description, boolean isDone, String dueDate) {
-            super(description, isDone);
-            this.due = parseDue(dueDate);
-        }
-
-        private static LocalDate parseDue(String s) {
-            try {
-                return LocalDate.parse(s, IO_FMT);
-            } catch (DateTimeParseException e) {
-                throw new IllegalArgumentException("Invalid deadline date '" + s + "'. Expected yyyy-MM-dd.", e);
-            }
-        }
-
-        @Override String getTaskType() { return "[D]"; }
-
-        @Override String getFullDescription() {
-            return getTaskType() + super.toString() + " (by: " + due.format(VIEW_FMT) + ")";
-        }
-
-        @Override public String toString() { return getFullDescription(); }
-
-        @Override String toFileString() {
-            return "D | " + (isDone ? "1" : "0") + " | " + description + " | " + due.format(IO_FMT);
-        }
-    }
-
-    // Event class
-    static class Event extends Task {
-        private static final DateTimeFormatter IO_FMT = DateTimeFormatter.ISO_LOCAL_DATE_TIME;       // e.g., 2015-02-20T06:30
-        private static final DateTimeFormatter VIEW_FMT = DateTimeFormatter.ofPattern("MMM d yyyy HH:mm");
-
-        private final LocalDateTime start;
-        private final LocalDateTime end;
-
-        // start/end must be ISO: yyyy-MM-dd'T'HH:mm[:ss]
-        Event(String description, String start, String end) {
-            super(description);
-            this.start = parseDateTime(start);
-            this.end = parseDateTime(end);
-            validateOrder();
-        }
-
-        Event(String description, boolean isDone, String start, String end) {
-            super(description, isDone);
-            this.start = parseDateTime(start);
-            this.end = parseDateTime(end);
-            validateOrder();
-        }
-
-        private static LocalDateTime parseDateTime(String s) {
-            try {
-                return LocalDateTime.parse(s, IO_FMT); // accepts both with & without seconds
-            } catch (DateTimeParseException ex) {
-                throw new IllegalArgumentException(
-                        "Invalid event datetime '" + s + "'. Expected ISO, e.g., 2015-02-20T06:30.", ex);
-            }
-        }
-
-        private void validateOrder() {
-            if (end.isBefore(start)) {
-                throw new IllegalArgumentException("Event end time cannot be before start time.");
-            }
-        }
-
-        @Override String getTaskType() { return "[E]"; }
-
-        @Override String getFullDescription() {
-            return getTaskType() + super.toString()
-                    + " (from: " + start.format(VIEW_FMT) + " to: " + end.format(VIEW_FMT) + ")";
-        }
-
-        @Override public String toString() { return getFullDescription(); }
-
-        @Override String toFileString() {
-            return "E | " + (isDone ? "1" : "0") + " | " + description + " | "
-                    + start.format(IO_FMT) + " | " + end.format(IO_FMT);
-        }
-    }
-
     public static void main(String[] args) {
-        ArrayList<Task> tasks = new ArrayList<>();
+        TaskList tasks = new TaskList();
         Scanner scanner = new Scanner(System.in);
 
         printLine();
@@ -271,7 +109,7 @@ public class Usagi {
         printLine();
     }
 
-    private static void displayTaskList(ArrayList<Task> tasks) {
+    private static void displayTaskList(TaskList tasks) {
         printLine();
         if (tasks.isEmpty()) {
             System.out.println("Your list is empty! Add some tasks first.");
@@ -284,7 +122,7 @@ public class Usagi {
         printLine();
     }
 
-    private static void addTodoTask(ArrayList<Task> tasks, String input) throws UsagiException {
+    private static void addTodoTask(TaskList tasks, String input) throws UsagiException {
         String description = input.substring(5).trim();
         if (description.isEmpty()) {
             throw new EmptyDescriptionException("todo");
@@ -299,7 +137,7 @@ public class Usagi {
         printLine();
     }
 
-    private static void addDeadlineTask(ArrayList<Task> tasks, String input) throws UsagiException {
+    private static void addDeadlineTask(TaskList tasks, String input) throws UsagiException {
         String withoutPrefix = input.substring(9).trim();
         if (withoutPrefix.isEmpty()) {
             throw new EmptyDescriptionException("deadline");
@@ -330,7 +168,7 @@ public class Usagi {
         printLine();
     }
 
-    private static void addEventTask(ArrayList<Task> tasks, String input) throws UsagiException {
+    private static void addEventTask(TaskList tasks, String input) throws UsagiException {
         String withoutPrefix = input.substring(6).trim();
         if (withoutPrefix.isEmpty()) {
             throw new EmptyDescriptionException("event");
@@ -368,7 +206,7 @@ public class Usagi {
         printLine();
     }
 
-    private static void handleMarkCommand(ArrayList<Task> tasks, String input, boolean markAsDone) throws UsagiException {
+    private static void handleMarkCommand(TaskList tasks, String input, boolean markAsDone) throws UsagiException {
         try {
             String[] parts = input.split(" ", 2);
             if (parts.length < 2) {
@@ -401,7 +239,7 @@ public class Usagi {
         }
     }
 
-    private static void deleteTask(ArrayList<Task> tasks, String input) throws UsagiException {
+    private static void deleteTask(TaskList tasks, String input) throws UsagiException {
         try {
             String[] parts = input.split(" ", 2);
             if (parts.length < 2) {
@@ -426,8 +264,8 @@ public class Usagi {
         }
     }
 
-    private static ArrayList<Task> readFileContent(String filePath) throws IOException {
-        ArrayList<Task> tasks = new ArrayList<>();
+    private static TaskList readFileContent(String filePath) throws IOException {
+        TaskList tasks = new TaskList();
 
         File f = new File(filePath);
         File folder = f.getParentFile();
@@ -480,10 +318,10 @@ public class Usagi {
         return tasks;
     }
 
-    private static void writeFileContent(ArrayList<Task> tasks, String filePath) throws IOException {
+    private static void writeFileContent(TaskList tasks, String filePath) throws IOException {
         FileWriter fw = new FileWriter(filePath);
-        for (Task t : tasks) {
-            fw.write(t.toFileString() + System.lineSeparator());
+        for (int i = 0; i < tasks.size(); i++) {
+            fw.write(tasks.get(i).toFileString() + System.lineSeparator());
         }
         fw.close();
     }
